Fix firmware build errors, add ESP-IDF build automation via build_idf.py
- Add missing #include "http_client.h" in main.c (implicit declaration error)
- Cast IP octets to (int) in web_config.c snprintf (format type mismatch)
- Exclude esp_lcd component in CMakeLists.txt (internal compiler error)
- Add build_idf.py to work around ESP-IDF MSYSTEM check in MSYS2/Git Bash
- Add 'firmware' make target to top-level Makefile
- Update .gitignore for firmware/build/ and firmware/sdkconfig

+++ firmware/main/web_config.c
@@ -250,7 +250,8 @@ static esp_err_t status_handler(httpd_req_t *req)
         wifi_mgr_get_info(ssid, &ip, &rssi);
         cJSON_AddStringToObject(root, "ssid", ssid);
         snprintf(ip_str, sizeof(ip_str), "%d.%d.%d.%d",
-                 ip & 0xFF, (ip >> 8) & 0xFF, (ip >> 16) & 0xFF, (ip >> 24) & 0xFF);
+                 (int)(ip & 0xFF), (int)((ip >> 8) & 0xFF),
+                 (int)((ip >> 16) & 0xFF), (int)((ip >> 24) & 0xFF));
         cJSON_AddStringToObject(root, "ip", ip_str);
         cJSON_AddNumberToObject(root, "rssi", rssi);
     }
